requirements.txt
streamlit>=1.24.0
streamlit-qrcode-scanner>=0.1.0
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
pytest>=7.3.1  # for testing
sqlalchemy>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
aiohttp>=3.11.0